requirements.txt
langchain-ollama
langchain-core
langchain-chroma
langgraph
requests
pyyaml
chromadb
python-dotenv
markdown
bs4
